Add files via upload
Mars has a sidereal year of 668.5907 Martian days. A Martian day is 24 hours 39 minutes, 35.244 seconds in length. In each decade there are 6 leap years that last 669 days and 4 normal years that last 668 days. Odd years, and years that are divisible by 10 are leap years. Create a program that calculates whether the input year is a Martian leap year or not.