Add files via upload
Develop an interactive version of a Celsius to Fahrenheit conversion program which prompts the user to input a value either in Celsius or in Fahrenheit. If a value in Celsius in input it will convert it to Fahrenheit and display the result. If a value in Fahrenheit in input it will convert the value to Celsius and display the result. The program should continue to prompt the user for new inputs. Typing "exit" or "Exit" should terminate. If the user does not enter a numeric value (other than "exit" or "Exit") the program should notify the user that "I don't understand. Try again."